Add session management
Session allow us to get session information out of MMStatus object.

I'm not so proud of the way  Session is implements as it force us to use
`Box <dyn SessionBase>` type for passing session (because we need a
match to have same type on both arms of match).

Perhaps an Enum would be easier to use.